requirements.txt
numpy==2.0.1
opencv_python==4.10.0.84
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
